Move cloud_api_helper under utils/ dir.
