+++ gslib/commands/cp.py
@@ -29,7 +29,6 @@ from gslib.cat_helper import CatHelper
 from gslib.command import Command
 from gslib.command_argument import CommandArgument
 from gslib.commands.compose import MAX_COMPONENT_COUNT
-from gslib.cloud_api import GetCloudApiInstance
 from gslib.copy_helper import CreateCopyHelperOpts
 from gslib.copy_helper import GetSourceFieldsNeededForCopy
 from gslib.copy_helper import GZIP_ALL_FILES
@@ -50,6 +49,7 @@ from gslib.storage_url import IsCloudSubdirPlaceholder
 from gslib.storage_url import StorageUrlFromString
 from gslib.third_party.storage_apitools import storage_v1_messages as apitools_messages
 from gslib.utils import parallelism_framework_util
+from gslib.utils.cloud_api_helper import GetCloudApiInstance
 from gslib.utils.constants import DEBUGLEVEL_DUMP_REQUESTS
 from gslib.utils.constants import NO_MAX
 from gslib.utils.posix_util import ConvertModeToBase8
